\name{mcmc2dataframe}
\alias{mcmc2dataframe}
\title{Convert Markov Chain Monte Carlo Objects to Dataframes}

\usage{
mcmc2dataframe(obj)
}

\arguments{
  \item{obj}{An mcmc object.}
}

\description{
  The function `mcmc2dataframe' is used to convert an mcmc object to a
  dataframe.
}
 
\value{
  A dataframe built from obj.
}

\keyword{models}

\seealso{
   \code{\link{mcmc2}}
}
