% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp-utils.R
\name{plotHDPChangepoint}
\alias{plotHDPChangepoint}
\title{Posterior Changepoint Probabilities from HDP-HMM}
\usage{
plotHDPChangepoint(
  mcmcout,
  main = "Posterior Changepoint Probabilities",
  xlab = "Time",
  ylab = "",
  start = 1
)
}
\arguments{
\item{mcmcout}{The \code{mcmc} object containing the posterior density
sample from a changepoint model. Note that this must be from a
  HDP-HMM sampler.}

\item{main}{Title of the plot}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{start}{The time of the first observation to be shown in the time
series plot.}
}
\description{
Plot the posterior density of regime change.
}
\seealso{
\code{\link{HDPHMMpoisson}}, \code{\link{HDPHMMnegbin}}, \code{\link{HDPHSMMnegbin}}
}
\keyword{hplot}
